requirements.txt
# ================================================================
# TOKO SEMBAKO MURAH JAYA
# Python Dependencies
# ================================================================
# 
# Install semua dependencies:
#   pip install -r requirements.txt
# 
# Update dependencies:
#   pip install --upgrade -r requirements.txt
# 
# ================================================================

# ----------------------------------------------------------------
# CORE FRAMEWORK
# ----------------------------------------------------------------
# Flask - Web framework utama
Flask==3.0.0

# Werkzeug - WSGI utilities & password hashing (Scrypt)
# Dependency dari Flask, tapi kita pin untuk konsistensi
Werkzeug==3.0.1

# Jinja2 - Template engine (dependency Flask)
Jinja2==3.1.2

# ----------------------------------------------------------------
# DATABASE
# ----------------------------------------------------------------
# PyMySQL - MySQL/MariaDB driver untuk Python
# Pure Python, tidak perlu kompilasi C extensions
PyMySQL==1.1.0

# Cryptography - Untuk PyMySQL SSL/TLS connections
# Diperlukan untuk koneksi database yang aman di production
cryptography==41.0.7

# ----------------------------------------------------------------
# ENVIRONMENT & CONFIGURATION
# ----------------------------------------------------------------
# python-dotenv - Load environment variables dari file .env
# Untuk menyimpan credentials dengan aman
python-dotenv==1.0.0

# ----------------------------------------------------------------
# PRODUCTION SERVER
# ----------------------------------------------------------------
# Gunicorn - Production-ready WSGI HTTP Server
# Digunakan untuk deployment ke Railway, Render, Heroku, dll
# Catatan: Hanya berjalan di Linux/Unix, tidak di Windows
gunicorn==21.2.0

# ----------------------------------------------------------------
# DEVELOPMENT DEPENDENCIES (Opsional)
# ----------------------------------------------------------------
# Uncomment jika diperlukan untuk development:
# 
# pytest==7.4.3           # Testing framework
# pytest-flask==1.3.0     # Flask testing utilities
# black==23.12.1          # Code formatter
# flake8==6.1.0           # Linter

# ================================================================
# CATATAN DEPLOYMENT
# ================================================================
# 
# PythonAnywhere:
#   - Tidak perlu gunicorn (sudah built-in)
#   - Pastikan versi Python >= 3.8
# 
# Railway/Render:
#   - Butuh gunicorn
#   - Butuh Procfile dengan: web: gunicorn app:app
# 
# Heroku:
#   - Butuh gunicorn
#   - Butuh Procfile dan runtime.txt
# 
# ================================================================
